chore(ui): consolidate debounced search fields (#9051)
* chore(ui): consolidate debounced search fields

* add debounced search

* cleanup

* force un-controlled

* cleanup import path
