+++ app/src/components/field/index.tsx
@@ -4,4 +4,5 @@ export * from "./CredentialField";
 export * from "./CredentialInput";
 export * from "./NumberField";
 export * from "./SearchIcon";
+export * from "./DebouncedSearch";
 export { TextArea } from "react-aria-components";
